move implementation to Walker trait
