+++ diamond-api/src/context.rs
@@ -1,0 +1,8 @@
+use crate::opts::*;
+use crate::storage::Walker;
+
+#[derive(Clone)]
+pub struct Context<T: Walker> {
+    pub args: Args,
+    pub walker: T,
+}
