+++ diamond-api/src/lib.rs
@@ -1,7 +1,11 @@
 pub mod application;
-pub mod error;
-pub mod find;
+pub mod context;
 pub mod opts;
-pub mod parse;
-pub mod render;
-pub mod render_target;
+pub mod storage;
+
+pub(crate) mod error;
+pub(crate) mod find;
+pub(crate) mod parse;
+pub(crate) mod render;
+pub(crate) mod render_target;
+pub(crate) mod test_utils;
